#13 inner class 추출 및 Getter 삭제, Enum class 한곳으로 정리
