+++ SyrupPayToken/Claims/Elements/Element.cs
@@ -1,0 +1,7 @@
+﻿namespace SyrupPayToken.Claims
+{
+    interface Element
+    {
+        void ValidRequired();
+    }
+}
