+++ SyrupPayToken/Claims/Elements/RegistrationRestrictions.cs
@@ -1,0 +1,17 @@
+﻿using Newtonsoft.Json;
+
+namespace SyrupPayToken.Claims
+{
+    [JsonObject(MemberSerialization.OptIn)]
+    public sealed class RegistrationRestrictions
+    {
+        [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
+        private string matchedUser;
+
+        public string MatchedUser
+        {
+            get { return matchedUser; }
+            set { matchedUser = value; }
+        }
+    }
+}
