+++ SyrupPayToken/SyrupPayToken.csproj
@@ -55,6 +55,23 @@
     <Compile Include="AbstractConfiguredTokenBuilder.cs" />
     <Compile Include="ClaimConfigurerAdapter.cs" />
     <Compile Include="Claims\AbstractTokenConfigurer.cs" />
+    <Compile Include="Claims\Elements\Accept.cs" />
+    <Compile Include="Claims\Elements\AdditionalDiscount.cs" />
+    <Compile Include="Claims\Elements\Bank.cs" />
+    <Compile Include="Claims\Elements\CardInstallmentInformation.cs" />
+    <Compile Include="Claims\Elements\Element.cs" />
+    <Compile Include="Claims\Elements\Enums.cs" />
+    <Compile Include="Claims\Elements\Error.cs" />
+    <Compile Include="Claims\Elements\ISOCode.cs" />
+    <Compile Include="Claims\Elements\Loyalty.cs" />
+    <Compile Include="Claims\Elements\MonthlyInstallment.cs" />
+    <Compile Include="Claims\Elements\Offer.cs" />
+    <Compile Include="Claims\Elements\PaymentInformationBySeller.cs" />
+    <Compile Include="Claims\Elements\PaymentRestriction.cs" />
+    <Compile Include="Claims\Elements\Plan.cs" />
+    <Compile Include="Claims\Elements\ProductDeliveryInfo.cs" />
+    <Compile Include="Claims\Elements\RegistrationRestrictions.cs" />
+    <Compile Include="Claims\Elements\ShippingAddress.cs" />
     <Compile Include="Claims\MapToSyrupPayUserConfigurer.cs" />
     <Compile Include="Claims\MerchantUserConfigurer.cs" />
     <Compile Include="Claims\OrderConfigurer.cs" />
